rider/requirements.txt
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
django-redis
drf-yasg
django-filter
channels
channels[daphne]
channels-redis
drf-nested-routers
confluent-kafka
geopy
requests